refactor(playground): move XMSS/LMS stateful signing from Classical to PQC tab
XMSS and LMS/HSS are PQC algorithms (NIST SP 800-208), but the working panels
lived under the Classical sign tab while the PQC sub-tab showed an "under
construction" stub. Move them to their correct home:

- extract the working XmssPanel/LmsPanel into hsm/StatefulHashSignPanels.tsx
- render them under the PQC -> "XMSS / LMS" sub-tab via an XMSS/LMS toggle,
  replacing the stub (the "requires softhsmv3 bump" note was stale — the
  engine already works in the current build)
- remove XMSS/LMS from the Classical tab (a move, not a duplicate)
